Update building and installation sections on README
Updates the README to include a section on building.  Also rewrites
the section on installation to indicate that no installer is
available and refer to the FAQ.

These changes are incomplete because the link to the FAQ is
missing (the installer explanation has not yet been added to
the FAQ).