Create build and install targets in top level SConstruct (#131, #123) (#175)
* Create build and install targets in top level SConstruct

This change allows the user to run a complete build
and to stage install files using a single command for each.
The old method of building and installing is still available,
but this method is simpler because it requires fewer commands
and doesn't require changing directories.

* Add quotes around installdir

Adds quotes around places where the installdir variable is
used so that the build and install scripts will work with
directories that contain spaces.

* Add install target back to SConstruct

Adds the install target to the SConstruct as an alias for the
stage_install target.  This allows users to continue building
and installing using the old commands.