Add symbolic link to docs
The docs need to be referenced from docs/geedocs so that they
will be available on opengee.org and from
earth_enterprise/docs/landing_page/manual so that they will be
available when GEE in installed.  This change adds a symlink from
manual to geedocs to avoid duplicating the docs.