Move documentation files to geedocs
Originally, geedocs was a symlink to
earth_enterprise/docs/landing_page/manual, which contained the
actual documentation.  However, jekyll (the tool that builds
opengee.org) has trouble with the symlink and throws an error.
To avoid this error, this commit moves the documentation to
geedocs and makes landing_page/manual a link to that, which
works fine for both the website and the help files installed
with fusion and server.