Eliminate NULL pointer error in date/time function (#451)
The hours, minutes, and seconds objects in QDateWrapper.cpp will
be NULL if the object is initialized with just a date and no
time.  This change eliminates a case in which the object might
try to set the time when the time objects do not exist.  This
bug sometimes manifested itself by causing a seg fault when
opening vector projects that were created by the run_tutorial
script.