Allow icons on map layers
This commit addresses multiple related things
1) Issue #881, which had errors returning to fusion users when saving a
map layer with an icon chosen for shields or points.  This was caused by
having no image decoders registered in the skia third-party library.

2) The skia library was not being built correctly by SCons. The source
for the skia library exists in a tarball, which is extracted as a step
in the build process, but SCons was trying to gather list of cpp files
up front before the files were extracted. Now the list of files is
gathered from the contents of the tarball.

3) To work around the issue above, the third-party libraries were being
built twice. This worked because on the second round, the cpp files
existed on disk at the right time. This second round is no longer
necessary.