individual deadlock scenarios taken care of by a quick rewrite, drastically simpler logic is in place now
the method of terminating cache should be examined. something is
preventing objects from being released, even after simplifying logic
down to simple adding and subtracting from a thread-safe object counter

its my assumption that these files are in use for a disproportionate amount of
time and/or not being released until the very end of usage, all at one time.

when adding in a timeout in code or bumping the purge percentage (i.e. the
percentage of which the cache is full before terminating) down to 0, the gauge
tests complete just fine.

it is my hypothesis that even with the class design, this issue will still
persist