add xml cache purge (#1174)
* added terminate code
terminates when cache is full
due to pointer semantics, its best to terminate on release of documents and parsers

* made some changes after testing

* added update to documentation

* made last minor change

* refactored portions into a helper class

* refactored to make helper class simpler

* unlock from parser creatiom

* logic change

* individual deadlock scenarios taken care of by a quick rewrite,
drastically simpler logic is in place now

the method of terminating cache should be examined. something is
preventing objects from being released, even after simplifying logic
down to simple adding and subtracting from a thread-safe object counter

its my assumption that these files are in use for a disproportionate amount of
time and/or not being released until the very end of usage, all at one time.

when adding in a timeout in code or bumping the purge percentage (i.e. the
percentage of which the cache is full before terminating) down to 0, the gauge
tests complete just fine.

it is my hypothesis that even with the class design, this issue will still
persist

* checking after passing tests, will clean up now

* cleaned up code

* for some reason the documentation related to xereces init params didn't get merged previously ???

* switched percentages over to levels

* locks on rw operations

* made changes to docs to reflect changes to purge levels

* update with guard taken off of reads

* refactored to get rid of purge lock, to allow for multiple documents to be processed
and to allow them to clear out when it is time to purge

* removed debugging statements

* added documentation

* added rw lock

* refactored into read-write lock

* Clean up locking

* reverted changes to reflect appropriate 5.2.5 documentation