Issue #280: Validate Pasted KML on Server
Currently pasted KML is trimmed and validated in Javascript. Yet
then the untrimmed KML is sent to the server, where it is
implicitly trusted by the process to build a globe.

This change maintains trimming of entered KML, and adds a server-
side check before writing standardized KML to the polygon file.