1598: Clean up RedHat and Centos build instructions
Removed dependency on wget, unless it's needed to run the gtest build script.

Added instructions for installing new versions of Git on EL7

Changed git2u-all to git216, to decrease the number of dependencies pulled in.

Moved Python 2.7 install into the "install the standard tools"
section, because python27 libraries are being installed there already.

Removed some packages that don't exist in EL6 repos, and aren't needed
on that platform.

Removed explicit installation of gfortran on EL6 because it's not
required.  Note that it gets installed anyway as part of the
devtoolset-2-toolchain package.