1774 Port Error Handler to WSGI (#1775)
* 1774 Port Error Handler to WSGI

Fixes #1774

The gehttpd_error_handler.conf currently redirects errors to a CGI solution. This requires a new process to be forked on every single 404, effectively causing gecutter to DDOS itself. This change ports the error handler to a WSGI Daemon process. This allows much more bandwidth for handling error pages and responses.